+++ index.html
@@ -226,7 +226,7 @@
 			</div>
 		</section>
 	</template>
-	<script src="/js/function.js"></script>
-	<script src="/js/main.ts"></script>
+	<script src="/js/function.ts"></script>
+	<script type="module" src="/js/main.ts"></script>
 </body>
 </html>
